+++ modules/airflow_storage/versions.tf
@@ -20,11 +20,11 @@ terraform {
 
     google = {
       source  = "hashicorp/google"
-      version = ">= 3.53, < 7"
+      version = ">= 3.53, < 8"
     }
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-composer:airflow_storage/v5.2.0"
+    module_name = "blueprints/terraform/terraform-google-composer:airflow_storage/v6.2.0"
   }
 }