+++ modules/create_environment_v2/versions.tf
@@ -20,21 +20,21 @@ terraform {
 
     google = {
       source  = "hashicorp/google"
-      version = ">= 5.14, < 7"
+      version = ">= 5.14, < 8"
     }
 
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = ">= 5.14, < 7"
+      version = ">= 5.14, < 8"
     }
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-composer:create_environment_v2/v5.2.0"
+    module_name = "blueprints/terraform/terraform-google-composer:create_environment_v2/v6.2.0"
   }
 
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-composer:create_environment_v2/v5.2.0"
+    module_name = "blueprints/terraform/terraform-google-composer:create_environment_v2/v6.2.0"
   }
 
 }