+++ test/setup/main.tf
@@ -16,7 +16,7 @@
 
 module "project" {
   source  = "terraform-google-modules/project-factory/google"
-  version = "~> 17.0"
+  version = "~> 18.0"
 
   name              = "ci-composer"
   random_project_id = "true"
@@ -32,7 +32,8 @@ module "project" {
     "dns.googleapis.com",
     "compute.googleapis.com",
     "iam.googleapis.com",
-    "container.googleapis.com"
+    "container.googleapis.com",
+    "cloudbuild.googleapis.com",
   ]
   activate_api_identities = [
     {
@@ -61,7 +62,8 @@ module "service_project" {
     "composer.googleapis.com",
     "compute.googleapis.com",
     "iam.googleapis.com",
-    "container.googleapis.com"
+    "container.googleapis.com",
+    "cloudbuild.googleapis.com",
   ]
   activate_api_identities = [
     {